Split code
The classes are divided into separate header files
